Fix Circular import issue
This commit fixes a circular import issue in certificates.
Ticket: #56256
