+++ gui/system/views.py
@@ -71,6 +71,7 @@ from freenasUI.middleware.client import client, CallTimeout, ClientException, Va
 from freenasUI.middleware.exceptions import MiddlewareError
 from freenasUI.middleware.form import handle_middleware_validation
 from freenasUI.middleware.notifier import notifier
+from freenasUI.middleware.util import get_validation_errors
 from freenasUI.middleware.zfs import zpool_list
 from freenasUI.network.models import GlobalConfiguration
 from freenasUI.storage.models import Volume
@@ -82,8 +83,7 @@ from freenasUI.system.utils import (
     debug_generate,
     factory_restore,
     run_updated,
-    is_update_applied,
-    certificate_common_post_create
+    is_update_applied
 )
 from middlewared.plugins.update import CheckUpdateHandler, get_changelog, parse_changelog
 
@@ -105,6 +105,44 @@ def _info_humanize(info):
     info['datetime'] = info['datetime'].replace(tzinfo=None)
     info['datetime'] = localtz.fromutc(info['datetime'])
     return info
+
+
+def certificate_common_post_create(action, request, **kwargs):
+
+    form, job, verrors, job_id = None, None, None, None
+
+    if request.session.get('certificate_create'):
+        form = getattr(
+            forms, request.session['certificate_create']['form']
+        )(request.session['certificate_create']['payload'], **kwargs)
+        job_id = request.session['certificate_create']['job_id']
+        form.is_valid()
+        form._middleware_action = action
+        verrors = get_validation_errors(job_id)
+        if verrors:
+            handle_middleware_validation(form, verrors)
+        with client as c:
+            job = c.call(
+                'core.get_jobs',
+                [['id', '=', job_id]],
+            )
+
+        del request.session['certificate_create']
+
+    if not job:
+        if job_id:
+            error = f'Job {job_id} does not exist'
+        else:
+            error = '"certificate_create" key does not exist in session'
+
+        job = {
+            'state': 'FAILED',
+            'error': error
+        }
+    else:
+        job = job[0]
+
+    return form, job, verrors
 
 
 def system_info(request):
